% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPH_Coltparkmeadow.R
\name{PPH_Coltparkmeadow}
\alias{PPH_Coltparkmeadow}
\title{An empirical tripartite pollinator-plant-herbivore network}
\usage{
data(PPH_Coltparkmeadow)
}
\description{
A pollinator-plant-herbivore tripartite network from Villa-Galaviz et. al. 2020. Journal of Animal Ecology
}
\details{
An 'igraph' object. This is an empirical tripartite network provided in Villa-Galaviz et. al. 2020. This network has three guilds of species: pollinators, plants and herbivores. Pollinators and plants form the subnetwork with mutualistic interactions, and plants and herbivores form the subnetwork with antagonistic interactions. No intra-guild interactions. Plants are the shared set of species.
}
\examples{
data(PPH_Coltparkmeadow)

}
\references{
Villa-Galaviz, E., S. M. Smart, E. L. Clare, S. E. Ward, and J. Memmott. 2021. Differential effects of fertilisers on pollination and parasitoid interaction networks. Journal of animal ecology 90:404-414.
}
\keyword{datasets}
