% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_motif.R
\name{Multi_motif}
\alias{Multi_motif}
\title{Defined multilayer motifs}
\usage{
Multi_motif(
  type = c("M111", "M112", "M113", "M211", "M212", "M213", "M311", "M312", "M121-1",
    "M122-1", "M122-2", "M122-3", "M123-1", "M123-2", "M123-3", "M123-4", "M123-5",
    "M221-1", "M221-2", "M221-3", "M222-1", "M222-2", "M222-3", "M222-4", "M222-5",
    "M222-6", "M222-7", "M222-8", "M321-1", "M321-2", "M321-3", "M321-4", "M321-5",
    "M131", "M132-1", "M132-2", "M132-3", "M132-4", "M132-5", "M231-1", "M231-2",
    "M231-3", "M231-4", "M231-5", "all")
)
}
\arguments{
\item{type}{Character. 45 values representing each and all types of multilayer motifs.}
}
\value{
If \code{type} is one of 44 motifs of definition, function returns a corresponding motif of "igraph".

If \code{type} is "all", function returns a list containing 44 corresponding motifs of "igraph".
}
\description{
Acquiring each or all of 44 motifs of definition of multilayer network.
}
\examples{

type<-c("M111","M112","M113","M211","M212","M213","M311","M312","M121-1",
"M122-1","M122-2","M122-3","M123-1","M123-2","M123-3","M123-4","M123-5",
"M221-1","M221-2","M221-3","M222-1","M222-2","M222-3","M222-4","M222-5",
"M222-6","M222-7","M222-8","M321-1","M321-2","M321-3","M321-4","M321-5",
"M131","M132-1","M132-2","M132-3","M132-4","M132-5","M231-1","M231-2",
"M231-3","M231-4","M231-5")

m<-Multi_motif("M222-4")
plot(m,main="M222-4")

mr <- par(mfrow=c(6,8),mar=c(1,1,3,1))
for(i in 1:44){
    plot(Multi_motif("all")[[i]],
    vertex.size=30, vertex.label=NA,
    vertex.color="blue",main=type[i])
}
par(mr)

}
