% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_icc.R
\name{node_icc}
\alias{node_icc}
\title{Interconnection centrality for connector nodes in a tripartite network}
\usage{
node_icc(network.or.subnet_mat1, subnet_mat2 = NULL, weighted = F)
}
\arguments{
\item{network.or.subnet_mat1}{Igraph object or matrix. An "igraph" object with node attribute 'level' or a matrix representing one subnetwork.}

\item{subnet_mat2}{The matrix representing one subnetwork.}

\item{weighted}{Logical. If true, link strengths of connector nodes are used. Default to FALSE.
#' @details
This function calculates interconnection degree, betweenness and closeness centrality for connector nodes.
For binary networks, interconnection degree centrality of each connector species is defined as the product of its degree values from two subnetworks,
interconnection betweenness centrality is defined by the number of shortest paths from a-nodes to c-nodes going through connector species,
and interconnection closeness centrality is defined by the inverse of the sum of distances from connector species to both a-nodes and c-nodes.
For weighted networks, interaction strengths are used in the calculation of weighted degree, shorest path, and distance.
\strong{network.or.subnet_mat1 and subnet_mat2}

\strong{weighted}

If the \code{weighted} = FALSE, the input network can be an "igraph" object or two matrices.If a weighted network is provided, it will be transformed to a binary network.
If the \code{weighted} = TRUE, the input network can only be two matrices.

\strong{network.or.subnet_mat1 and subnet_mat2}

There are two types of inputs \code{network.or.subnet_mat1} that can be processed:
\itemize{
\item Input is a network of type "igraph" alone.
\item Must be entered as matrix with \code{subnet_mat2}.
}

If the inputs are two matrices, please make sure the rows of
\code{network.or.subnet_mat1} and \code{subnet_mat2} both represent the groups of connector species,i.e, the b-group species. If both matrices have row names, then the function matches row
names to produce connector nodes. Otherwise, row numbers are assigned as row names.}
}
\value{
Return a data frame with interconnection degree, betweenness and closeness.
}
\description{
Calculating interconnection centrality (degree, betweenness and closeness) specified for connector nodes in a tripartite network.
}
\examples{

## generate a random tripartite network
set.seed(12)
Net <- build_net(11,15,16,0.2)

data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow
node_icc(Net)
set.seed(13)
E(Net)$weight<-runif(length(E(Net)),0.1,1)#random weights assigned
node_icc(Net,weighted=T)

##input as binary matrices,with row names.
set.seed(12)
md1 <- matrix(sample(c(0,1),5*8,replace=TRUE),5,8,dimnames = list(paste0("b",1:5),paste0("c",1:8)))
md2 <- matrix(sample(c(0,1),20*30,replace=TRUE),20,30,dimnames = list(paste0("b",1:20),paste0("a",1:30)))
node_icc(md1,md2)

##input as weighted matrices,with row numbers as row names.
set.seed(17)
mdw1 <- matrix(sample(c(rep(0,20),runif(20,0,1))),5,8)
mdw2 <- matrix(sample(c(rep(0,500),runif(100,0,1))),20,30)
node_icc(mdw1,mdw2)
node_icc(mdw1,mdw2,weighted=T)

}
\references{
De Domenico, M., Nicosia, V., Arenas, A., & Latora, V. (2015). Structural
reducibility of multilayer networks. Nature communications, 6(1), 6864.

De Domenico, M., SolC)-Ribalta, A., Omodei, E., GC3mez, S., & Arenas, A.
(2013). Centrality in interconnected multilayer networks. arXiv preprint
arXiv:1311.2906.

De Domenico, M. (2022). Multilayer Networks: Analysis and Visualization.
Introduction to muxViz with R. Cham: Springer.

Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The pagerank citation
ranking: Bringing order to the web.

Magnani, M., Micenkova, B., & Rossi, L. (2013). Combinatorial analysis of
multiple networks. arXiv preprint arXiv:1303.4986.
}
