% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc.R
\name{poc}
\alias{poc}
\title{Proportion of connector nodes}
\usage{
poc(network.or.subnet_mat1, subnet_mat2 = NULL)
}
\arguments{
\item{network.or.subnet_mat1}{Igraph or matrix. Both weighted or unweighted networks are accepted. A tripartite network of 'igraph' class which contains three groups (a,b,c) of species and two subnetworks (P including a and b groups of species, Q including b and c groups of species). If matrices are provided, this should be the matrix representing subnetwork P that have links between a- and b-group species.
Within this matrix, each row and column represents b-group species and a-group species respectively. For matrix, row names are required, otherwise row number is used.
Elements of matrix are non-zero values if two species are linked, and 0 otherwise.}

\item{subnet_mat2}{The matrix representing subnetwork Q that have links between b- and c-group species.
Within this matrix, each row and column represents b-group species and c-group species respectively.
Elements of matrix are non-zero values if two species are linked, and 0 otherwise.If \code{network.or.subnet_mat1} is "igraph", \code{subnet_mat2} defaults to NULL.}
}
\value{
Print a "POC= ;" and Return a numeric value representing the proportion of connector species in shared group of species.
}
\description{
Calculating the proportion of connector nodes in the shared set of nodes.
}
\details{
\strong{network.or.subnet_mat1 and subnet_mat2}

There are two types of inputs \code{network.or.subnet_mat1} that can be processed:
\itemize{
\item Input is a network of type "igraph" alone.
\item Must be entered as matrix with \code{subnet_mat2}.
}

If the inputs are two matrices, please make sure the rows of
\code{network.or.subnet_mat1} and \code{subnet_mat2} both represent the groups of connector species,i.e, the b-group species. If both matrices have row names, then the function matches row
names to produce connector nodes. Otherwise, row numbers are assigned as row names.
}
\examples{

## generate a random tripartite network
set.seed(12)
Net <- build_net(11,15,16,0.2)

data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow
poc(Net)

##input as binary matrices,with row names.
md1 <- matrix(sample(c(0,1),8*11,replace=TRUE),8,11,dimnames = list(paste0("b",1:8),paste0("c",1:11)))
md2 <- matrix(sample(c(0,1),10*12,replace=TRUE),10,12,dimnames = list(paste0("b",1:10),paste0("a",1:12)))
poc(md1,md2)

##input as weighted matrices,with row numbers as row names.
mdw1 <- matrix(sample(c(rep(0,40),runif(48,0,1))),8,11)
mdw2 <- matrix(sample(c(rep(0,40),runif(80,0,1))),10,12)
poc(mdw1,mdw2)


}
\references{
Battiston, F., Nicosia, V. & Latora, V. (2014) Structural measures for multiplex networks. Physical Review E, 89, 032804.

Domínguez-García, V., & Kéfi, S. (2024). The structure and robustness of ecological networks with two interaction types. PLOS Computational Biology, 20(1), e1011770.

Guimera, R. & Amaral, L.A.N. (2005) Cartography of complex networks: modules and universal roles. Journal of Statistical Mechanics: Theory and Experiment, 2005, P02001.
}
