% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPH_Coltparkmeadow.R
\name{PPH_Coltparkmeadow}
\alias{PPH_Coltparkmeadow}
\title{A pollinator-plant-herbivore network from Villa-Galaviz et. al. 2020. Journal of Animal Ecology}
\usage{
data(PPH_Coltparkmeadow)
}
\description{
A pollinator-plant-herbivore network from Villa-Galaviz et. al. 2020. Journal of Animal Ecology
}
\details{
A list with "igraph". This is a classical example of multilayer interaction network.
}
\examples{
data(PPH_Coltparkmeadow)

}
\references{
Villa‐Galaviz, E., S. M. Smart, E. L. Clare, S. E. Ward, and J. Memmott. 2021. Differential effects of fertilisers on pollination and parasitoid interaction networks. Journal of animal ecology 90:404-414.
}
\keyword{datasets}
