% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPH_Coltparkmeadow.R
\name{PPH_Coltparkmeadow}
\alias{PPH_Coltparkmeadow}
\alias{data}
\title{A pollinator-plant-herbivore network from Villa-Galaviz et. al. 2020. Journal of Animal Ecology}
\format{
A data frame with
}
\usage{
data(PPH_Coltparkmeadow)
}
\description{
A pollinator-plant-herbivore network from Villa-Galaviz et. al. 2020. Journal of Animal Ecology
}
\examples{
data(PPH_Coltparkmeadow)

}
